WIP: Move random test utils to cypress, configure available folders to test by jest
