+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "dev-front": "webpack-dev-server --mode development",
     "dev-back": "nodemon back-server.js",
-    "utest": "jest src",
-    "utest-cover": "jest src --coverage",
+    "utest": "jest src && jest cypress/utils",
+    "utest-cover": "jest src --coverage && jest cypress/utils --coverage",
     "ftest": "cypress run",
     "ftest-spec": "cypress run --spec",
     "ftest-ui": "cypress open",
